misc:  Change shift expressions to unsigned types. Shifting signed values to a result that is not representable has undefined behaviour.

+++ src/server/sv_client.c
@@ -108,7 +108,7 @@ void SV_GetChallenge(netadr_t from)
 		// this is the first time this client has asked for a challenge
 		challenge = &svs.challenges[oldest];
 
-		challenge->challenge = ((rand() << 16) ^ rand()) ^ svs.time;
+		challenge->challenge = (((unsigned int) rand() << 16) ^ (unsigned int)rand()) ^ svs.time;
 		challenge->adr       = from;
 		challenge->firstTime = svs.time;
 		challenge->firstPing = 0;
